Split resolver type reference validation
